feat - temperatures min & max

+++ WeatherZapto.Data.Services/ISupervisor/ISupervisorTemperature.cs
@@ -2,8 +2,8 @@
 {
     public interface ISupervisorTemperature
     {
-        Task<IEnumerable<float>> GetTemperatures(string location, DateTime day);
-        Task<float?> GetTemperatureMin(string location, DateTime day);
-        Task<float?> GetTemperatureMax(string location, DateTime day);
+        Task<IEnumerable<double>> GetTemperatures(string location, DateTime day);
+        Task<double?> GetTemperatureMin(string location, DateTime day);
+        Task<double?> GetTemperatureMax(string location, DateTime day);
     }
 }
